requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
gunicorn==21.2.0
google-generativeai==0.1.0
PyMuPDF==1.22.5
python-pptx==0.6.21
Pillow==10.0.1
openpyxl==3.1.3
pytesseract==0.3.10
